+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ## BC breaks
  - protected postWrite method not called anymore from BaseConnectorBundle\\Writer\\Doctrine\\Writer.
  If you need it, override the write method, call the parent and add your code after.
+ - constructor of Pim\Bundle\BaseConnectorBundle\Writer\Doctrine\Writer has changed
+ - Pim\Bundle\TransformBundle\Cache\ProductCacheClearer has been renamed Pim\Bundle\TransformBundle\Cache\CacheClearer
 
 # 1.2.0